miner, jsoncpp->rapidjson (web monitor)
Rapidjson offers very accurate control... I'm not sure it buys me
anything real but the API is indeed well designed and thought out. I
like the fact you cannot create stuff by accident and the no-conversions
rule... but this is a bit the opposite of what JSON is supposed to be
("easygoing").
Keeping rapidjson for the time being.