446-minor
It was my understanding real open source should have no names attached
(according to an old presentation by google I remember). Apparently the
author is an important exception. I still want to be very relaxed on
those things so I stamp my name to kernels only which are possibly the
most important part of the project.