This refactor introduces 1) a new way to generate work units and 2) corrects a bug with difficulty management
== 1. Work units ==
In legacy miners, work units are a pretty huge entity including everything needed to generate header to hash. The same more or less happened here and it had more or less the same issues. The new method involves AbstractWorkSource creating an object derived from AbstractWorkFactory. It is still a bit ugly but allows to hide quite some details in AbstractWorkSource itself. Various stuff moved around to cope with this change.

== 2. Work unit change and difficulty ==
Stratum really manages two streams of data: the data to hash and the difficulty setting at which the data is to be hashed. When M8M was first written the idea was to generate a new WU every time the difficulty changed but I forgot enforcing that rule. Albeit this simplifies things, it just makes no sense.
Now the miner can receive difficulty changes independently from new blocks of data to hash.
Note the previous behavior resulted in short streams of rejected shares when diff was increased. This in turn caused the server to receive below target shares and rollback to lower diff.
For high-hashrate miners this might as well never reach a stable state. It seems the amount of rejects is indeed lower.