793 - Correct two newly discovered bugs.
The first problem in mainHelpers.h is a string not being unicode.
The second problem in network.cpp is more subtle. In certain occasions, it is possible for a server to be reachable through multiple routes. In this case, connection failed as soon as one candidate failed, which is not what I want.